Merge 1st para of "Why important" with page intro
Per discussion of EOWG 5/23, here are draft edits that delete the first paragraph of the "Why is this important?" section -- merging the concepts into the initial paragraphs of the Table Concepts page. If you merge this in, would you specifically call out a few people to review this? I'm a bit nervous making these edits as this seems so fundamental.  Also, I'll work on the next sentence of "Why is this important" in a separate proposed file change. Thanks!