Text-only note
I am concerned by this as it is not clear that a text-only version would not meet all the situations suggested above. I am not saying text-only versions are not bad, but I think these reasons are not necessarily that compelling and could easily be used as supporting arguments for a text only version.

I might go so far as to suggest removing this note... at the least, some minor edits.