Unclear statement
I don't know that it is clear what 'non-screen readers' is referring to. Is this assistive technologies that are not screen readers? It might be beneficial to include an example. So, 'At the time of writing this tutorial, this approach is not fully supported by all web browsers and assistive technologies, for example Braille displays.