Suggestion may lead to missing alt attributes
This may be me over worrying but I was slightly concerned that saying 'there is no need to repeat the information in the text alternative...' may result in some simply removing the alt attribute in similar cases. The example clarifies this slightly but since there is no code example it could be misunderstood or ignored.

I wonder if this note might be better presented in Tips and Tricks? The first part is sufficient to ensure that author input is required, the second part is supporting info that might even be applicable in other situations.