Typo double check
I'm seeing in this text that "remember" is spelled correctly.  It's not spelled correctly in the github version I'm looking at as of 9:40 CST Feb. 19 (http://w3c.github.io/wai-tutorials/images/complex/) -- so this is mostly to be sure it really is going to get corrected.  While I was at it I suggest changing a comma to a dash in that same sentence.