Double check on sentence fragment
What I'm seeing on the github rendition is"  "The long description provides detailed information, including scales, values, relationships and trends that are represented visually. For example, to point out the declining values for site 1, consistent values for site 2, and increasing values for site 3 that are encoded in the bar chart." The second sentence is a fragment but appears to be corrected in this version.  That's the most important thing.  Meanwhile, suggest changing verb "should" in this version.