Significant copy rewrite.
I rewrote this page to have a higher readability score using Flesch-Kincaid. Many individual sentences of the original, current version has a readability score of less than 38, but fortunately an overall score of 53.2. This made the grade level of readability to be roughly the equivalent of 11th grade US education (not saying anything about our education system, here). To put this in perspective, Readers Digest has an average readability score of 65. Noting that the widest audience could be using this, I wanted to get it to be roughly on an 8th grade level if possible.

My final edits achieved a readability of 63.6 with a roughly 8.6 grade level (probably a remedial US freshman high school grade). I hope this helps.