Permalink z-index causes overlap problems
This only really occurs when there are two permalinks relatively close together... admittedly not going to happen often, but I noticed it when I was testing. If you open a sharebox immediately above (in the page) another Share link, then the Sharelink overlaps the sharebox. It happens because the z-index on the .permalink container takes priority over the z-index of the contained .sharebox. Not sure why the .permalink needs the z-index; the .sharebox certainly does.