Change label id to for attribute
I was testing a couple of the examples on this page with a screen reader, and noticed that there wasn't a label being read for the select example. I changed the id attribute to a for attribute to associate the label with the select so that it will be read on screen reader focus.